Proof-of-concept runtime patching for hooks.
Avoids the need to edit the binary to enable
hooks, instead editing dinput_hooks.lst is enough.
Also avoids the complexity and performance impact
of the map and trampolines needed by the previous approach.
Converting hooks with arguments will be a bit more work,
but will allow making them proper functions arguemnts instead
of a pointer to an array.