Add Titlebar Option for window drag
Allows an option in the hook_windowed.cfg file for adding a titlebar to the window style.  This allows the user to drag the window, minimize, maximize and restore.  It is finicky however and some more work may need to be done for lost or gain focus events to deal with the mouse hook functionality integration.